Fetch song data and begin work on SongDetail component.
